Retry connecting to Redis in Server
